replace old test_gru_layer_parallel with the new one
